requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2